% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpmsplit-package.R
\docType{package}
\name{dpmsplit-package}
\alias{dpmsplit}
\alias{dpmsplit-package}
\title{dpmsplit: Disaggregate Inflows or Outflows}
\description{
Disaggregate estimated inflows to or outflows from a population, using data on component flows. The sensitivity to different data sources can be controlled by 'alterability' parameters. Based on methods proposed by Quenneville and Rancourt.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ONSdigital/dpmsplit}
  \item Report bugs at \url{https://github.com/ONSdigital/dpmsplit/issues/new?assignees=&labels=bug&projects=&template=bug_report_template.md}
}

}
\author{
\strong{Maintainer}: Office for National Statistics \email{demographic.methods@ons.gov.uk}

Authors:
\itemize{
  \item John Bryant
}

Other contributors:
\itemize{
  \item Rhosanna Ellum [contributor]
  \item Aidan Metcalfe [contributor]
  \item Andrew Taylor [contributor]
  \item Daniel Ward [contributor]
}

}
\keyword{internal}
