% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{sim_reported}
\alias{sim_reported}
\title{Simulated reported outflows, by type}
\format{
A data frame with 38 rows and 3 columns:
\describe{
  \item{age}{Five-year age group.}
  \item{time}{2021 or 2022.}
  \item{internal}{Reported migration to other regions
    within the country.}
  \item{external}{Reported migration to other countries.}
}
}
\usage{
sim_reported
}
\description{
Simulated outflows, distinguishing "internal"
flows (ie to other regions within the country)
from "external" flows (ie to other countries).
}
\keyword{datasets}
