% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{sim_reported_em}
\alias{sim_reported_em}
\title{Simulated reported flows for external emigration}
\format{
A data frame with 76 rows and 5 columns:
\describe{
  \item{age}{Five-year age group.}
  \item{time}{2021 or 2022.}
  \item{region_orig}{a or b.}
  \item{region_dest}{x.}
  \item{count}{Numeric column containing outflow values.}
}
}
\usage{
sim_reported_em
}
\description{
Simulated outflows from two regions in the
same country to one external region, mimicking
reported external emigration data.
}
\keyword{datasets}
