% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{sim_reported_im}
\alias{sim_reported_im}
\title{Simulated reported flows for external immigration}
\format{
A data frame with 76 rows and 5 columns:
\describe{
  \item{age}{Five-year age group.}
  \item{time}{2021 or 2022.}
  \item{region_orig}{x.}
  \item{region_dest}{a or b.}
  \item{count}{Numeric column containing inflow values.}
}
}
\usage{
sim_reported_im
}
\description{
Simulated inflows from one external region to
two regions in the same country, mimicking reported
external immigration data.
}
\keyword{datasets}
