% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{sim_reported_int}
\alias{sim_reported_int}
\title{Simulated reported flows for internal migration}
\format{
A data frame with 76 rows and 5 columns:
\describe{
  \item{age}{Five-year age group.}
  \item{time}{2021 or 2022.}
  \item{region_orig}{a or b.}
  \item{region_dest}{a or b.}
  \item{count}{Numeric column containing migration values.}
}
}
\usage{
sim_reported_int
}
\description{
Simulated flows between two regions, mimicking
reported internal migration data.
}
\keyword{datasets}
