% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{sim_totals}
\alias{sim_totals}
\title{Simulated estimates for total outflows}
\format{
A data frame with 38 rows and 3 columns:
\describe{
  \item{age}{Five-year age group.}
  \item{time}{2021 or 2022.}
  \item{count}{A list column, containing numeric vectors of length 10}
}
}
\usage{
sim_totals
}
\description{
Simulated estimates of outflows from a population,
mimicking the results from running the base model,
though with only 10 interations.
}
\keyword{datasets}
