% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-docs.R
\docType{data}
\name{sim_totals_out}
\alias{sim_totals_out}
\title{Simulated estimates for total outflows for multiple populations}
\format{
A data frame with 76 rows and 4 columns:
\describe{
  \item{age}{Five-year age group.}
  \item{time}{2021 or 2022.}
  \item{region}{a or b.}
  \item{count}{A list column, containing numeric vectors of length 10}
}
}
\usage{
sim_totals_out
}
\description{
Simulated estimates of outflows from multiple
populations, mimicking the results from running
the base model, though with only 10 iterations.
}
\keyword{datasets}
