% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split_single}
\alias{split_single}
\title{Split flows for single population}
\usage{
split_single(totals, reported, alter = NULL, zeros_to_ones = FALSE)
}
\arguments{
\item{totals}{Marginal totals.
A data frame with columns of 'classification
variables', plus outcome variable \code{"count"}, in any order.
The \code{"count"} column can be a numeric vector, or can be
a list of numeric vectors of equal length.}

\item{reported}{Reported migration counts.
A data frame with a subset of the classification
variables from \code{totals}, and, instead of \code{"count"},
two or more columns of reported migration flows.}

\item{alter}{Alterability values. A named
list of data frames. The data frames specify
alterability values for each type of reported migration
flow. The data frames each have a subset of the classification
variables from \code{reported}, plus a column
called \code{"alter"}. If no data frame is supplied for a
migration flow, then every level of that flow is assumed to have
an alterability value of 1. If no value for \code{alter}
is supplied, then all alterability values are set to 1.}

\item{zeros_to_ones}{If a row of \code{reported}
consists entirely of zeros, replace it with a row of ones.
Defaults to \code{FALSE}.}
}
\value{
A data frame with the same classification variables
as \code{totals}, but with the \code{"count"} column split
into separate flows.
}
\description{
Given combined flows into or out of a single population,
split the flows into separate streams.
}
\details{
The \code{totals} argument holds the combined flows
plus one or more classification variables.
For instance, \code{totals} might describe
all outflows from a population, disaggregated by age,
sex/gender, cohort, and time. The flows are held
in a variable called \code{"count"}.
The count variable can be an ordinary numeric vector,
or, if the flows are uncertain, a list column.
If \code{totals} was generated from the base model,
then the count variable will be a list column.

The \code{reported} argument holds data on
the disaggregated flows, plus
one or more classification variables.
For instance, \code{reported} might have data on
(i) migration to other regions of the country,
held in a variable called \code{"internal_out"},
and (ii) migration to other countries, held in a
variable called \code{"emigration"}, as well as
classification variables \code{"age"} and \code{"time"}.
\code{reported} does not need to
have all the classification variables in \code{totals},
but if it does have a variable, then that variable
must have all the levels that are found in the
\code{totals} version.

The \code{alter} argument holds values for the
'alterability' parameter, which is applied
to the data in \code{reported}. \code{alter} is optional:
if no value is supplied, then all observations in
\code{reported} are assigned an alterability score of 1.
\code{alter} is a named list of numbers
or data frames. Each number or
data frame refers to a type of flow in
\code{reported}. If a flow in \code{"reported"} does not have a
corresponding data frame in \code{"alter"}, then
all values for that flow receive an alterability
score of 1. If a flow in \code{"reported"} has a
number in \code{"alter"} (eg \code{1.7}), then
that number is used as the alterability score for
all values in the flow.
If a flow has a data frame, then different elements
within the flow receive different alterability scores
(found in the \code{alter} value column of the data frame.)
As with \code{report}, \code{alter} does not have to have all
the classification variables that are found in \code{totals},
but if it does have a variable, that variable must have
a complete set of levels.

The returned value is identical to \code{totals},
except that the \code{"count"} column is replaced by
the flow columns for \code{"reported"}, scaled so that
the flows add up to the values in \code{"count"}.
Values with high alterability scores are modified more
than values with low alterability scores.

The adjustment process is based on [Quenneville and Rancourt 2005](https://ec.europa.eu/eurostat/documents/3888793/5837169/KS-DT-05-016-EN.PDF.pdf/9f326f78-ca08-4344-a3ef-5b1f2cb515c9?t=1414779260000). Sometimes, if reported values
are being scaled downwards and a reported value is near zero but
has a high alterability score, the original Quenneville and  Rancourt
can produce negative values for flows. \code{split_single}
avoids negative values by setting the flow to zero and recalculating:
see the vignette to this package for details.

In small populations, sometimes all values in a
row of \code{reported} are zero while a corresponding
value in \code{totals} is non-zero, and the usual formula
breaks down. A simple solution is to
change all values in the row of \code{reported} to 1,
which implicitly weights each data source equally.
To apply this solution, set \code{zeros_to_ones} to
\code{TRUE}. If zero cells are common, a more sophisticated
approach, such as smoothing counts in \code{reported}
before passing it to function \code{split_single}, might
be appropriate.

\code{NA}s are permitted in \code{totals} and \code{reported},
and will propagate through to the results. \code{NA}s are not
permitted in \code{alter}.
}
\examples{
totals <- data.frame(
  age = c("20-24", "25-29", "20-24", "25-29"),
  sex = c("Female", "Female", "Male", "Male"),
  count = c(20, 24, 15, 8)
)
reported <- data.frame(
  age = c("20-24", "25-29", "20-24", "25-29"),
  sex = c("Female", "Female", "Male", "Male"),
  internal = c(10, 15, 12, 3),
  cross_border = c(7, 8, 5, 0),
  international = c(3, 5, 1, 4)
)
alter <- list(
  cross_border = 1.2,
  international = data.frame(
    age = c("20-24", "25-29"),
    alter = c(2, 2.5)
  )
)

split_single(
  totals = totals,
  reported = reported,
  alter = alter
)

}
\seealso{
\code{\link{split_multi}}
}
